refactor: rename class name 各ゲームごとにクラスが生成されていたので、まとめた
