+++ typescript-jest/src/ITennisGame.ts
@@ -1,0 +1,4 @@
+export interface ITennisGame {
+  wonPoint(playerName: string): void;
+  getScore(): string;
+}
