+++ typescript-jest/src/index.ts
@@ -1,4 +1,2 @@
+export { ITennisGame } from './ITennisGame';
 export { TennisGame } from './TennisGame';
-export { TennisGame1 } from './TennisGame1';
-export { TennisGame2 } from './TennisGame2';
-export { TennisGame3 } from './TennisGame3';
