+++ typescript-jest/tests/TennisTest.spec.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { TennisGame, TennisGame1, TennisGame2, TennisGame3 } from '../src';
+import { ITennisGame, TennisGame } from '../src';
 
 function getAllScores(): Array<[number, number, string]> {
   const testCases = path.resolve(__dirname, 'scores.json');
@@ -11,7 +11,7 @@ function getAllScores(): Array<[number, number, string]> {
 
 const scores: Array<[number, number, string]> = getAllScores();
 
-function checkScore(game: TennisGame, player1Score: number, player2Score: number, expectedScore: string): void {
+function checkScore(game: ITennisGame, player1Score: number, player2Score: number, expectedScore: string): void {
   const highestScore: number = Math.max(player1Score, player2Score);
   for (let i = 0; i < highestScore; i++) {
     if (i < player1Score) {
@@ -29,7 +29,7 @@ describe('TennisGame', () => {
   describe('TennisGame1', () => {
     scores.forEach(([player1Score, player2Score, expectedScore]) => {
       it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, () => {
-        checkScore(new TennisGame1('player1', 'player2'), player1Score, player2Score, expectedScore);
+        checkScore(new TennisGame(1, 'player1', 'player2'), player1Score, player2Score, expectedScore);
       });
     });
   });
@@ -37,7 +37,7 @@ describe('TennisGame', () => {
   describe('TennisGame2', () => {
     scores.forEach(([player1Score, player2Score, expectedScore]) => {
       it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, () => {
-        checkScore(new TennisGame2('player1', 'player2'), player1Score, player2Score, expectedScore);
+        checkScore(new TennisGame(2, 'player1', 'player2'), player1Score, player2Score, expectedScore);
       });
     });
   });
@@ -45,7 +45,7 @@ describe('TennisGame', () => {
   describe('TennisGame3', () => {
     scores.forEach(([player1Score, player2Score, expectedScore]) => {
       it(`scores ${player1Score}:${player2Score} as ${expectedScore}`, () => {
-        checkScore(new TennisGame3('player1', 'player2'), player1Score, player2Score, expectedScore);
+        checkScore(new TennisGame(3, 'player1', 'player2'), player1Score, player2Score, expectedScore);
       });
     });
   });
